third_party/WebKit/LayoutTests/editing/inserting/insert-after-delete-001.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => assert_selection(
    '<div contenteditable>^foo|</div>',
    selection => {
        selection.document.execCommand('delete');
        selection.document.execCommand('insertText', false, 'bar');
    },
    '<div contenteditable>bar|</div>'),
    'insertText after delete');
</script>
